tests/requirements.txt
mkdocs
mkdocs-material
mkdocstrings
systemrdl-compiler
invoke
numpy
pytest